requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.6.1
joblib==1.3.2
python-owasp-zap-v2.4
reportlab
